// src/integrations/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Database = {
  public: {
    Tables: {
// ...
      sops: {
        Row: {
          category: Database["public"]["Enums"]["sop_category"];
          created_at: string | null;
          description: string | null;
          generated_content: string | null;
          id: string;
          tags: string[] | null;
          title: string;
          updated_at: string | null;
          user_id: string;
          workflow_data: Json | null;
        };
        Insert: {
          category: Database["public"]["Enums"]["sop_category"];
          created_at?: string | null;
          description?: string | null;
          generated_content?: string | null;
          id?: string;
          tags?: string[] | null;
          title: string;
          updated_at?: string | null;
          user_id: string;
          workflow_data?: Json | null;
        };
        Update: {
          category?: Database["public"]["Enums"]["sop_category"];
          created_at?: string | null;
          description?: string | null;
          generated_content?: string | null;
          id?: string;
          tags?: string[] | null;
          title?: string;
          updated_at?: string | null;
          user_id?: string;
          workflow_data?: Json | null;
        };
        Relationships: [];
      };
// ...
    };
// ...
  };
};

type DefaultSchema = Database[Extract<keyof Database, "public">];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof Database },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof Database;
  }
    ? keyof Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never
> = DefaultSchemaEnumNameOrOptions extends { schema: keyof Database }
  ? Database[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
  ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
  : never;